Adiciona método main em Application
