+++ jogo-oito/src/main/java/Application.java
@@ -1,0 +1,12 @@
+import builder.JogoDosOitoBuilder;
+
+public class Application {
+
+    public static void main(String[] args) {
+        new JogoDosOitoBuilder()
+                .createButtons()
+                .configureMenu()
+                .configureInterface()
+                .build();
+    }
+}
